requirements.txt
APScheduler==3.6.1
astroid==2.3.0
autopep8==1.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pylint==2.4.0
python-dotenv==0.10.3
pytz==2019.2
requests==2.22.0
selenium==3.141.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.5
wrapt==1.11.2
